Added consteval function to compute needed permutation arrays of contraction info.
contraction info speciefies legs to be contracted with positive integers and open legs with negative integers. E.g.: [1,2,-3,-4]x[-1,1,2,-2] for a contraction of two rank 4 tensors.
Since this uses c++20 features consteval and nontrivial non-type-template-parameters NTTP it is guarded by macro and only available when compiling with a recent enough compiler.