Added new template paramater `TRank` to `CTM`.
This will control the amount of legs between the T- and A-tensors.
For conventional quantum states, there are 2 legs (bra and ket).
For classical paritition functions there is 1 leg.
For PEPO overlaps there would be 3 legs.
Another possibility is the merging of e.g. bra and ket leg to be in the TRank=1 case.