Improved the handling of observables.
Added two template parameters: 1) `typename Scalar`, 2) `bool HERMITIAN`.
The first describes the scalar type of the operator's matrix elements.
The second describes if the operator is Hermitian. That should be known at compile time to decide the return type of expectation values for this operator (complex for non-Hermitian, real for Hermitian)