New template parameter `ALL_OUT_LEGS` for `iPEPS` and `CTM`.
It controls whether the convention for the legs of the iPEPS are all in or two in/ two out as before.
This also determines the direction of legs of environment tensors.
The all in convention allows for C4v rotational symmetric tensors and Hermitian corner and edge tensors.
So this will allow a more effiecient CTM scheme if the symmetry is present.