Decomposed the method `tSVD` into two parts.
The first part is `full_svd_impl` which performs a svd without truncating small singular values. The second part is `cutoff_matrices` which performs the truncation of U, Sigma and Vdag for small singular values.
This is useful for code readability and also for AD.

Added static method `WeightedIdentity` which returns an identity tensor with the degeneracy of the block quantum-number as weights.