Added `warmup_chis` options for gradient optimization.
This runs the CTM at all chis specified in `warmup_chis` in nontracking mode and reuses the environment from smaller chis.
After that the CTM is started with the main chi value in nontracking mode and if this is converged, the tracked steps for backwards ad are performed.
An empty `warmup_chis` corresponds to the previous behavior.